Make --time-sec a float instead of int

+++ lib/core/optiondict.py
@@ -104,7 +104,7 @@ optDict = {
 
     "Techniques": {
         "tech": "string",
-        "timeSec": "integer",
+        "timeSec": "float",
         "uCols": "string",
         "uChar": "string",
         "uFrom": "string",
